requirements.txt
streamlit
langchain
langchain-community
google-generativeai
chromadb
sentence-transformers
python-dotenv
PyPDF2
